Update movement class

+++ Assets/Scripts/Movement.cs
@@ -2,6 +2,6 @@
     public float CordX { get; set; }
     public float CordZ { get; set; }
     public float Rotation { get; set; }
-    public float InputX { get; set; }
-    public float InputZ { get; set; }
+    public sbyte InputX { get; set; }
+    public sbyte InputZ { get; set; }
 }
